Added permission checks to the project and project task domain logic layer.

+++ src/dssgmkt/domain/proj.py
@@ -13,6 +13,7 @@ from ..models.common import (
 from django.db.models import Case, When
 
 from .common import validate_consistent_keys
+from dssgmkt.authorization.common import ensure_user_has_permission
 
 def filter_public_projects(query_set):
     return query_set.exclude(status=ProjectStatus.DRAFT) \
@@ -76,7 +77,12 @@ class ProjectService():
         return user.is_authenticated and (ProjectRole.objects.filter(project=proj, user=user).exists() or ProjectService.user_is_project_volunteer_official(user, proj))
 
     @staticmethod
+    def user_is_project_commenter(user, proj):
+        return user.is_authenticated and (ProjectService.user_is_project_member(user, proj) or ProjectService.user_is_project_volunteer(user, proj))
+
+    @staticmethod
     def get_project_changes(request_user, proj):
+        ensure_user_has_permission(request_user, proj, 'project.log_view')
         ## verify user permissions
         return ProjectLog.objects.filter(project = proj).order_by('-change_date')
 
@@ -88,6 +94,7 @@ class ProjectService():
     def add_project_comment(request_user, projid, project_comment):
         project = Project.objects.get(pk=projid)
         if project:
+            ensure_user_has_permission(request_user, project, 'project.comment_add')
             project_comment.project = project
             project_comment.author = request_user
             try:
@@ -100,12 +107,14 @@ class ProjectService():
     @staticmethod
     def save_project(request_user, projid, project):
         validate_consistent_keys(project, ('id', projid))
+        ensure_user_has_permission(request_user, project, 'project.information_edit')
         project.save()
 
     # TODO check for permissions (user is admin role of the organization in question)
     @staticmethod
     def add_staff_member(request_user, projid, project_role):
         project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.staff_edit')
         if project:
             project_role.project = project
             try:
@@ -123,6 +132,7 @@ class ProjectService():
     @staticmethod
     def save_project_role(request_user, projid, project_role):
         validate_consistent_keys(project_role, 'Role does not match project', (['project', 'id'], projid))
+        ensure_user_has_permission(request_user, project_role.project, 'project.staff_edit')
         project_role.save()
         # NotificationService.add_user_notification(organization_role.user,
         #                                             "Your role within " + organization_role.organization.name + " has been changed to " + organization_role.get_role_display() + ".",
@@ -133,6 +143,7 @@ class ProjectService():
     @staticmethod
     def delete_project_role(request_user, projid, project_role):
         validate_consistent_keys(project_role, 'Role does not match project', (['project', 'id'], projid))
+        ensure_user_has_permission(request_user, project_role.project, 'project.staff_remove')
         project_role.delete()
         # NotificationService.add_user_notification(organization_role.user,
         #                                             "Your role within " + organization_role.organization.name + " has been changed to " + organization_role.get_role_display() + ".",
@@ -142,14 +153,20 @@ class ProjectService():
 
     @staticmethod
     def get_all_project_staff(request_user, projid):
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.staff_view')
         return ProjectRole.objects.filter(project=projid).order_by('role')
 
     @staticmethod
     def get_all_project_volunteers(request_user, projid):
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.volunteers_view')
         return ProjectTaskRole.objects.filter(task__project__id=projid)
 
     @staticmethod
     def get_all_volunteer_applications(request_user, projid):
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.volunteers_view')
         return VolunteerApplication.objects.filter(task__project__id = projid).order_by(
                         Case(When(status=ReviewStatus.NEW, then=0),
                              When(status=ReviewStatus.ACCEPTED, then=1),
@@ -176,7 +193,8 @@ class ProjectTaskService():
         return ProjectTask.objects.get(pk=taskid, project=projid)
 
     @staticmethod
-    def get_all_tasks(request_user, proj): # TODO check that the user has permissions to take a look at all the tasks
+    def get_all_tasks(request_user, proj):
+        ensure_user_has_permission(request_user, proj, 'project.tasks_view')
         return ProjectTask.objects.filter(project=proj).order_by('estimated_start_date')
 
     @staticmethod
@@ -189,7 +207,7 @@ class ProjectTaskService():
         return ProjectTask.objects.filter(project=proj).exclude(stage=TaskStatus.COMPLETED).exclude(stage=TaskStatus.DELETED).order_by('estimated_start_date')
 
     @staticmethod
-    def get_volunteer_current_tasks(request_user, volunteer, projid):
+    def get_volunteer_current_tasks(request_user, volunteer, projid): ## TODO check user is the same as volunteer or user is project owner ?
         return ProjectTask.objects.filter(project__pk=projid,
                                           projecttaskrole__user=volunteer,
                                           stage__in=[TaskStatus.STARTED, TaskStatus.WAITING_REVIEW])
@@ -215,8 +233,7 @@ class ProjectTaskService():
         return ProjectTaskRole.objects.filter(task=taskid, role=TaskRole.VOLUNTEER).exists()
 
     @staticmethod
-    def save_task(request_user, projid, taskid, project_task):
-        validate_consistent_keys(project_task, 'Task not found in that project', ('id', taskid), (['project', 'id'], projid))
+    def save_task_internal(request_user, projid, taskid, project_task):
         project_task.save()
         # TODO calculate the project status correctly based on all the tasks
         # project_task.project.status = ProjectStatus.WAITING_REVIEW
@@ -224,16 +241,23 @@ class ProjectTaskService():
          # TODO move this to a separate method that modifies tasks (so effects are passed on to the project as needed)
 
     @staticmethod
+    def save_task(request_user, projid, taskid, project_task):
+        validate_consistent_keys(project_task, 'Task not found in that project', ('id', taskid), (['project', 'id'], projid))
+        ensure_user_has_permission(request_user, project_task.project, 'project.task_edit')
+        ProjectTaskService.save_task_internal(request_user, projid, taskid, project_task)
+
+    @staticmethod
     def mark_task_as_completed(request_user, projid, taskid, project_task_review):
         project_task = ProjectTask.objects.get(pk=taskid, project__id=projid)
         project = Project.objects.get(pk=projid)
         if project and project_task:
+            ensure_user_has_permission(request_user, project_task, 'project.volunteer_task_finish')
             with transaction.atomic():
                 project_task_review.task = project_task
                 project_task_review.review_result = ReviewStatus.NEW
                 project_task_review.save()
                 project_task.stage = TaskStatus.WAITING_REVIEW
-                ProjectTaskService.save_task(request_user, projid, taskid, project_task)
+                ProjectTaskService.save_task_internal(request_user, projid, taskid, project_task)
         else:
             if not project:
                 raise KeyError('Project not found ' + str(projid))
@@ -243,6 +267,7 @@ class ProjectTaskService():
     @staticmethod
     def delete_task(request_user, projid, project_task):
         validate_consistent_keys(project_task, 'Task not found in that project', (['project', 'id'], projid))
+        ensure_user_has_permission(request_user, project_task.project, 'project.task_delete')
         project_task.delete()
         # TODO calculate the project status correctly based on all the tasks
         # project_task.project.status = ProjectStatus.WAITING_REVIEW
@@ -250,12 +275,13 @@ class ProjectTaskService():
          # TODO move this to a separate method that modifies tasks (so effects are passed on to the project as needed)
 
     @staticmethod
-    def create_default_task(request_user, projid): # TODO check permissions of the user within this project
+    def create_default_task(request_user, projid):
         # TODO calculate the project status correctly based on all the tasks
         # project_task.project.status = ProjectStatus.WAITING_REVIEW
         # project_task.project.save()
         # TODO move this to a separate method that modifies tasks (so effects are passed on to the project as needed)
         project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.task_edit')
         if project:
             project_task = ProjectTask()
             project_task.name = 'New project task'
@@ -274,6 +300,8 @@ class ProjectTaskService():
 
     @staticmethod
     def get_project_task_review(request_user, projid, taskid, reviewid):
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project_task.project, 'project.task_review_view')
         task_review = ProjectTaskReview.objects.get(pk=reviewid)
         validate_consistent_keys(task_review, 'Task review not found in that project', (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
         return task_review
@@ -283,6 +311,7 @@ class ProjectTaskService():
         validate_consistent_keys(task_review, 'Task review not found in that project', (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
         project_task = ProjectTask.objects.get(pk=taskid)
         project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project_task.project, 'project.task_review_do')
         with transaction.atomic(): # TODO check that there are no other active volunteers in this task before marking the task as completed? Some data model improvements are needed
             task_review.save()
             if task_review.review_result == ReviewStatus.ACCEPTED:
@@ -297,6 +326,7 @@ class ProjectTaskService():
     @staticmethod
     def accept_task_review(request_user, projid, taskid, task_review): # TODO check that the review request is in status NEW
         validate_consistent_keys(task_review, 'Task review not found in that project', (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
+        ensure_user_has_permission(request_user, task_review.task.project, 'project.task_review_do')
         task_review.review_result = ReviewStatus.ACCEPTED
         task_review.review_date = timezone.now()
         ProjectTaskService.save_task_review(request_user, projid, taskid, task_review)
@@ -309,6 +339,7 @@ class ProjectTaskService():
     @staticmethod
     def reject_task_review(request_user, projid, taskid, task_review): # TODO check that the review request is in status NEW
         validate_consistent_keys(task_review, 'Task review not found in that project', (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
+        ensure_user_has_permission(request_user, task_review.task.project, 'project.task_review_do')
         task_review.review_result = ReviewStatus.REJECTED
         task_review.review_date = timezone.now()
         ProjectTaskService.save_task_review(request_user, projid, taskid, task_review)
@@ -321,6 +352,7 @@ class ProjectTaskService():
     @staticmethod
     def cancel_volunteering(request_user, projid, taskid, project_task_role):
         validate_consistent_keys(project_task_role, 'Task role not found in that project', (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
+        ensure_user_has_permission(request_user, project_task_role.task, 'project.volunteer_task_cancel')
         if project_task_role.user != request_user:
             raise ValueError('Role does not match current user')
         else:
@@ -330,7 +362,7 @@ class ProjectTaskService():
                 if not ProjectTaskService.task_has_volunteers(request_user, taskid):
                     project_task.stage = TaskStatus.STARTED ## or not started?
                     project_task.accepting_volunteers = True
-                    ProjectTaskService.save_task(request_user, projid, taskid, project_task)
+                    ProjectTaskService.save_task_internal(request_user, projid, taskid, project_task)
 
                 # NotificationService.add_user_notification(request_user,
                 #                                             "You left " + organization_role.organization.name,
@@ -346,7 +378,8 @@ class ProjectTaskService():
 
 
     @staticmethod
-    def apply_to_volunteer(request_user, projid, taskid, task_application_request):
+    def apply_to_volunteer(request_user, projid, taskid, task_application_request): # TODO check the user is not already accepted for the task
+        # TODO check the user has a volunteer profile
         project_task = ProjectTask.objects.get(pk=taskid, project__id=projid)
         validate_consistent_keys(project_task, 'Task not found in that project', (['project', 'id'], projid))
         task_application_request.status = ReviewStatus.NEW
@@ -356,18 +389,21 @@ class ProjectTaskService():
 
     @staticmethod
     def get_volunteer_application(request_user, projid, taskid, volunteer_application_pk):
+        volunteer_application = VolunteerApplication.objects.get(pk=volunteer_application_pk, task__id=taskid, task__project__id=projid)
+        ensure_user_has_permission(request_user, volunteer_application, 'project.volunteers_application_view')
         ## We can avoid doing this by using all the constraints in the DB query
         # volunteer_application = VolunteerApplication.objects.get(pk=volunteer_application_pk)
         # validate_consistent_keys(volunteer_application, (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
         # return volunteer_application
         ## ... like so
-        return VolunteerApplication.objects.get(pk=volunteer_application_pk, task__id=taskid, task__project__id=projid)
+        return volunteer_application # TODO fix comment right above
 
     @staticmethod
     def save_volunteer_application(request_user, projid, taskid, volunteer_application):
         validate_consistent_keys(volunteer_application, (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
         project_task = ProjectTask.objects.get(pk=taskid)
         project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.volunteers_application_review')
         with transaction.atomic():
             if not volunteer_application.is_new():
                 volunteer_application.resolution_date = timezone.now()
@@ -379,6 +415,7 @@ class ProjectTaskService():
                 task_role.user = volunteer_application.volunteer
                 task_role.save()
 
+    # TODO remove all the permission validation in accept/reject methods that delegate on a save method? Document it clearly
     @staticmethod
     def accept_volunteer(request_user, projid, taskid, volunteer_application): # TODO check that the review request is in status NEW
         validate_consistent_keys(volunteer_application, (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
@@ -406,23 +443,30 @@ class ProjectTaskService():
 
     @staticmethod
     def get_project_task_requirements(request_user, projid, taskid):
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.task_requirements_view')
         return ProjectTaskRequirement.objects.filter(task=taskid, task__project__id=projid)
 
     @staticmethod
     def add_task_requirement(request_user, projid, taskid, requirement):
         project_task = ProjectTask.objects.get(pk=taskid)
         validate_consistent_keys(project_task, (['project', 'id'], projid))
+        ensure_user_has_permission(request_user, project_task.project, 'project.task_requirements_edit')
         requirement.task = project_task
         requirement.save()
 
     @staticmethod
     def save_task_requirement(request_user, projid, taskid, requirement):
         validate_consistent_keys(requirement, (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.task_requirements_edit')
         requirement.save()
 
     @staticmethod
     def delete_task_requirement(request_user, projid, taskid, requirement):
         validate_consistent_keys(requirement, (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.task_requirements_delete')
         requirement.delete()
 
     @staticmethod
@@ -437,6 +481,8 @@ class ProjectTaskService():
     def save_project_task_role(request_user, projid, taskid, project_task_role):
         # Do not check the task ID because we are changing it, so it does not match
         validate_consistent_keys(project_task_role, (['task', 'project', 'id'], projid))
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.volunteers_edit')
         project_task_role.save()
         # NotificationService.add_user_notification(organization_role.user,
         #                                             "Your role within " + organization_role.organization.name + " has been changed to " + organization_role.get_role_display() + ".",
@@ -447,6 +493,8 @@ class ProjectTaskService():
     @staticmethod
     def delete_project_task_role(request_user, projid, taskid, project_task_role):
         validate_consistent_keys(project_task_role, (['task', 'id'], taskid), (['task', 'project', 'id'], projid))
+        project = Project.objects.get(pk=projid)
+        ensure_user_has_permission(request_user, project, 'project.volunteers_remove')
         project_task_role.delete()
         # NotificationService.add_user_notification(organization_role.user,
         #                                             "Your role within " + organization_role.organization.name + " has been changed to " + organization_role.get_role_display() + ".",
